% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Lab 7 fire data}
\format{
## `fire`
A data frame with 16 rows and 2 columns:
\describe{
  \item{DISTANCE}{DAMAGE}
  ...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
Lab 7 fire data
}
\keyword{datasets}
