% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{a positive integer}

\item{n}{a positive integer}

\item{p}{a number between 0 and 1}
}
\value{
A barplot of the number of successes of a binomial distrubution repeated iter times
}
\description{
mybin
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
