% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of resamples taken}

\item{x}{a sample to bootstrap from}

\item{fun}{what statistic to find from the data}

\item{alpha}{the level of confidence}

\item{cx}{text size}

\item{...}{}
}
\value{
A graph of the sample results and the confidence interval for the statistic.
}
\description{
myboot2
}
\examples{
myboot2(10000,x=rnorm(25,mean=25,sd=10),fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
