% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{number of values in each binomial simulation}

\item{iter}{number of binomial simulations done}

\item{p}{probability of a success for the binomial distribution}

\item{...}{additional parameters}
}
\value{
a histogram of the mean values from the binomial distributions
}
\description{
mycltb
}
\examples{
mycltb(n=4,iter=10000,p=0.3)
}
