% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{an integer}

\item{sigma}{an integer}

\item{a}{an integer}
}
\value{
a plot of the density with the region less than a shaded and the probability of the shaded region.
}
\description{
myncurve
}
\examples{
myncurve(5,1,5)
}
