% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{A positive integer representing the number of seats available}

\item{gamma}{a number between 0 and 1 representing the probability of overbooking}

\item{p}{a number between 0 and 1 representing the probability of a ticket buyer showing up}
}
\value{
two plots of the discrete and continuous solutions and a list with the solutions
}
\description{
ntickets
}
\examples{
n = ntickets(400,0.02,0.95)
}
